VPC related tests moved to new file
